refactor(cmd/registry): move registry subcommands from cmd to internal/registry
