+++ cmd/registry.go
@@ -19,6 +19,10 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"github.com/falcosecurity/falcoctl/internal/registry/login"
+	"github.com/falcosecurity/falcoctl/internal/registry/logout"
+	"github.com/falcosecurity/falcoctl/internal/registry/pull"
+	"github.com/falcosecurity/falcoctl/internal/registry/push"
 	commonoptions "github.com/falcosecurity/falcoctl/pkg/options"
 )
 
@@ -31,10 +35,10 @@ func NewRegistryCmd(ctx context.Context, opt *commonoptions.CommonOptions) *cobr
 		Long:                  "Interact with OCI registries",
 	}
 
-	cmd.AddCommand(NewLoginCmd(ctx, opt))
-	cmd.AddCommand(NewLogoutCmd(opt))
-	cmd.AddCommand(NewPushCmd(ctx, opt))
-	cmd.AddCommand(NewPullCmd(ctx, opt))
+	cmd.AddCommand(login.NewLoginCmd(ctx, opt))
+	cmd.AddCommand(logout.NewLogoutCmd(opt))
+	cmd.AddCommand(push.NewPushCmd(ctx, opt))
+	cmd.AddCommand(pull.NewPullCmd(ctx, opt))
 
 	return cmd
 }
