+++ internal/registry/logout/logout.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cmd
+package logout
 
 import (
 	"github.com/spf13/cobra"
